Large refactor to ensure we don't do cyclomatic calls
