+++ src/repositories/indexable-repository.php
@@ -9,6 +9,7 @@ use Yoast\WP\Lib\ORM;
 use Yoast\WP\SEO\Builders\Indexable_Builder;
 use Yoast\WP\SEO\Helpers\Current_Page_Helper;
 use Yoast\WP\SEO\Helpers\Indexable_Helper;
+use Yoast\WP\SEO\Helpers\Permalink_Helper;
 use Yoast\WP\SEO\Loggers\Logger;
 use Yoast\WP\SEO\Models\Indexable;
 
@@ -60,6 +61,13 @@ class Indexable_Repository {
 	protected $indexable_helper;
 
 	/**
+	 * Represents the permalink helper.
+	 *
+	 * @var Permalink_Helper
+	 */
+	protected $permalink_helper;
+
+	/**
 	 * Returns the instance of this class constructed through the ORM Wrapper.
 	 *
 	 * @param Indexable_Builder              $builder              The indexable builder.
@@ -67,7 +75,7 @@ class Indexable_Repository {
 	 * @param Logger                         $logger               The logger.
 	 * @param Indexable_Hierarchy_Repository $hierarchy_repository The hierarchy repository.
 	 * @param wpdb                           $wpdb                 The WordPress database instance.
-	 * @param Indexable_Helper               $indexable_helper     The indexable helper.
+	 * @param Permalink_Helper               $permalink_helper     The permalink helper.
 	 */
 	public function __construct(
 		Indexable_Builder $builder,
@@ -75,14 +83,14 @@ class Indexable_Repository {
 		Logger $logger,
 		Indexable_Hierarchy_Repository $hierarchy_repository,
 		wpdb $wpdb,
-		Indexable_Helper $indexable_helper
+		Permalink_Helper $permalink_helper
 	) {
 		$this->builder              = $builder;
 		$this->current_page         = $current_page;
 		$this->logger               = $logger;
 		$this->hierarchy_repository = $hierarchy_repository;
 		$this->wpdb                 = $wpdb;
-		$this->indexable_helper     = $indexable_helper;
+		$this->permalink_helper     = $permalink_helper;
 	}
 
 	/**
@@ -465,7 +473,7 @@ class Indexable_Repository {
 	 */
 	protected function ensure_permalink( $indexable ) {
 		if ( $indexable && $indexable->permalink === null ) {
-			$indexable->permalink = $this->indexable_helper->get_permalink_for_indexable( $indexable );
+			$indexable->permalink = $this->permalink_helper->get_permalink_for_indexable( $indexable );
 
 			// Only save if changed.
 			if ( $indexable->permalink !== null ) {
@@ -493,5 +501,21 @@ class Indexable_Repository {
 		$indexable_ids = $this->hierarchy_repository->find_children( $indexable );
 
 		return $this->find_by_ids( $indexable_ids );
+	}
+
+	public function reset_permalink( $type = null, $subtype = null ) {
+		$query = $this->query()
+			->set( 'permalink', null )
+			->set( 'permalink_hash', null );
+
+		if ( $type !== null ) {
+			$query->where( 'object_type', $type );
+		}
+
+		if ( $type !== null && $subtype !== null ) {
+			$query->where( 'object_sub_type', $subtype );
+		}
+
+		return $query->update_many();
 	}
 }
